+++ src/protocolgame.h
@@ -152,7 +152,7 @@ private:
 	void sendIcons(uint16_t icons);
 	void sendFYIBox(std::string_view message);
 
-	void sendDistanceShoot(const Position& from, const Position& to, uint8_t type);
+	void sendDistanceShoot(const Position& from, const Position& to, uint16_t type);
 	void sendMagicEffect(const Position& pos, uint16_t type);
 	void sendCreatureHealth(const Creature* creature);
 	void sendSkills();
